minor gui improvements : - smaller plugin name field, indented according plugin depth - C editor icon - IEC editor icon

+++ images/genicons.sh
@@ -4,9 +4,13 @@ INKSCAPE=inkscape
 
 for i in `cat icons.svg |grep -o -e '%%[^%]*%%'|sed 's/%//g'` 
 do
- echo "$INKSCAPE" icons.svg -z -e $i.png -i $i
- rm  -f $i.png
- "$INKSCAPE" icons.svg -z -e $i.png -i $i
+ if [ $i.png -nt icons.svg ]; then
+ 	echo "Skip $i"
+ else
+	rm  -f $i.png
+	echo "$INKSCAPE" icons.svg -z -e $i.png -i $i
+	"$INKSCAPE" icons.svg -z -e $i.png -i $i
+ fi
 done
 
 cp ico24.png brz.png
